requirements.txt
certifi==2025.6.15
cloudscraper==1.2.71
cryptography==42.0.8
customtkinter==5.2.2
darkdetect==0.8.0
dnspython==2.7.0
icmplib==3.0.4
impacket==0.12.0
ldap3==2.9.1
ldapdomaindump==0.10.0
maxminddb==2.7.0
pillow==11.3.0
psutil==7.0.0
pycryptodomex==3.23.0
pyOpenSSL==24.0.0
PyRoxy @ git+https://github.com/MatrixTM/PyRoxy.git@ea0f88dbc0573292ba5672124f69e4e7a31b544d
PySocks==1.7.1
requests==2.32.4
requests-toolbelt==1.0.0
yarl==1.20.1
